Update proxy config to work with jammy based proxy servers
The older proxy code was unable to successfully connect to bosh
through a jumpbox deployed with jammy stemcells.

We've removed the old manual proxy in favor of the socks5
proxy built into the bosh cli. The bosh cli library will
parse the BOSH_ALL_PROXY settings on load though, so we
implmeneted a function in bosh-utils to reload the proxy settings.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>